+++ src/main/java/com/rcc/dev/backend/repository/MapDepartmentRoleRepository.java
@@ -1,6 +1,0 @@
-package com.rcc.dev.backend.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MapDepartmentRoleRepository extends JpaRepository<MapDepartmentRole, Long> {
-}